fix id type in oauth_access_grants

+++ db/schema.rb
@@ -1168,7 +1168,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_06_17_142856) do
 
   create_table "oauth_access_grants", force: :cascade do |t|
     t.integer "resource_owner_id", null: false
-    t.integer "application_id", null: false
+    t.bigint "application_id", null: false
     t.string "token", null: false
     t.integer "expires_in", null: false
     t.text "redirect_uri", null: false
@@ -1184,7 +1184,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_06_17_142856) do
 
   create_table "oauth_access_tokens", force: :cascade do |t|
     t.integer "resource_owner_id"
-    t.integer "application_id", null: false
+    t.bigint "application_id", null: false
     t.string "token", null: false
     t.string "refresh_token"
     t.integer "expires_in"
